Fixed merge conflicts and moved ScheduleActivity
