+++ app/src/main/res/layout/fragment_activity_schedule.xml
@@ -5,7 +5,7 @@
     android:id="@+id/constraintLayout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".activity_schedule$DayFragment">
+    tools:context=".view.ScheduleActivity$DayFragment">
 
     <android.support.v7.widget.RecyclerView
         android:id="@+id/activity_schedule_recyclerview"
